requirements.txt
torch>=2.5.0
transformers>=4.46.0,<5.0
datasets>=3.0.0
accelerate>=1.0.0
deepspeed>=0.15.0
bitsandbytes>=0.43.0
openrlhf>=0.9.0
vllm>=0.6.0
docker>=7.0.0
tree-sitter>=0.23.0
pyyaml>=6.0
wandb>=0.18.0
tqdm>=4.66.0
numpy>=1.26.0
sentencepiece>=0.2.0
tiktoken>=0.7.0
tokenizers>=0.20.0
huggingface-hub>=0.25.0
flash-attn>=2.6.0
peft>=0.13.0
human-eval>=1.0
